+++ src/builder/mod.rs
@@ -25,6 +25,8 @@ pub(crate) fn build_file(pair: Pair<Rule>) -> Result<File> {
 
                         #[cfg(not(feature = "module-imports"))]
                         {
+                            use crate::error::NbclError;
+
                             return Err(NbclError::Ast {
                                 message: "module imports are disabled".into(),
                                 hint: Some(
@@ -67,6 +69,8 @@ pub(crate) fn build_file(pair: Pair<Rule>) -> Result<File> {
                     }
                     #[cfg_attr(not(feature = "lib-imports"), allow(unreachable_code))]
                     Rule::import_lib_stmt => {
+                        use crate::error::NbclError;
+
                         let mut inner = child.clone().into_inner();
                         let library_pair = inner.next().unwrap();