+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Not checking child count on native nodes.
 - Not throwing error when properties were passed to a component that expected none.
 - Wasm using outdated API.
+- No default feature build failing.
 
 ## [0.3.4] - 2026-05-21